Split test_cdemodels into multiple files
